Astroshaper.jl

A package for dynamical simulation of an asteroid.

Astroshaper.CERES — Constant

Ceres

Astroshaper.DIDYMOS — Constant

Asteroid 65803 Didymos (1996 GT)

Physical parameters

:GM : GM

:M : Mass

:μ : Standard gravitational parameter about Sun and Ryugu

Orbital elements

:a : Semi-mojor axis [AU]

:e : Eccentricity [-]

:I : Inclination [deg]

:Ω : Longitude of the ascending node [deg]

:ω : Argument of periapsis [deg]

:Φ : # Mean anomaly [deg]

Spin parameters

:α : Right ascension (RA) in equatorial coordinate system [deg]

:δ : Declination (Dec) in equatorial coordinate system [deg]

:P : Rotation period [h]

References

Small Body Database Lookup: https://ssd.jpl.nasa.gov/tools/sbdb_lookup.html#/?sstr=didymos

Naidu et al. (2020)

Astroshaper.DIMORPHOS — Constant

Dimorphos

:a : Semi-major axis of the mutual orbit with Dimorphos [m]

:P : Rotation periof of the mutual orbit with Dimorphos [h]

Astroshaper.EARTH — Constant

Earth

Astroshaper.ERIS — Constant

Eris

Astroshaper.JUPITER — Constant

Jupiter

Astroshaper.MARS — Constant

Mars

Astroshaper.MERCURY — Constant

Mercury

Astroshaper.MOON — Constant

Moon

Astroshaper.NEPTUNE — Constant

Neptune

Astroshaper.PLUTO — Constant

Pluto

Astroshaper.RYUGU — Constant

Asteroid 162173 Ryugu

Physical parameters

:GM : GM

:M : Mass

:μ : Standard gravitational parameter about Sun and Ryugu

Orbital elements

:a : Semi-mojor axis [AU]

:e : Eccentricity [-]

:I : Inclination [deg]

:Ω : Longitude of the ascending node [deg]

:ω : Argument of periapsis [deg]

:Φ : # Mean anomaly [deg]

Spin parameters

:α : Right ascension (RA) in equatorial coordinate system [deg]

:δ : Declination (Dec) in equatorial coordinate system [deg]

:P : Rotation period [h]

Astroshaper.SATURN — Constant

Saturn

Astroshaper.URANUS — Constant

Uranus

Astroshaper.VENUS — Constant

Venus

Astroshaper.Binary — Type

mutable struct Binary{T1, T2, T3, T4}

Describe the state of a binary asteroid

Fields

shape1 : Shape model of the primary

shape2 : Shape model of the secondary

orbit : Orbital elements of the primary

mutual_orbit : Mutual orbit parametes

spin1 : Spin parameters of the primary

spin2 : Spin parameters of the secondary

Astroshaper.Facet — Type

struct Face{T1, T2, T3, T4, T5, T6}

Triangular surface facet of a polyhedral shape model.

Note that the mesh normal indicates outward the polyhedron.

Fields

A::T1 : Position of 1st vertex

B::T1 : Position of 2nd vertex

C::T1 : Position of 3rd vertex

center::T1 : Position of mesh center

normal::T1 : Normal vector to mesh

area ::T2 : Area of mesh

visiblefacets::T3 : 1-D array of VisibleFacet

flux ::T4 : Energy flux from surrounding facets

temps ::T5 : Temperature profile in depth direction

_temps_ ::T5 : Pre-allocated vector for updating temperature profile

force ::T6 : Photon recoil force

Astroshaper.Flux — Type

mutable struct Flux{T}

Energy flux to a facet

Fields

sun ::T : Flux of solar radiation, F_sun

scat::T : Flux of scattered sunlight, F_scat

rad ::T : Flux of thermal radiation, F_rad

Astroshaper.HermiteParticle — Type

Particle{T1 <: AbstractVector, T2 <: Real} <: AbstractParticle

Paticle for N-body simulation (4th-degree Hermite integrator)

Fields

r # Position

v # Velocity

a # Acceleration

a¹ # 1st dervative of acceleration (jerk)

a² # 2nd dervative of acceleration

a³ # 3rd dervative of acceleration

⁺r # Position at next timestep

⁺v # Velocity at next timestep

⁺a # Acceleration at next timestep

⁺a¹ # 1st dervative of acceleration at next timestep

⁺a² # 2nd dervative of acceleration at next timestep

⁺a³ # 3rd dervative of acceleration at next timestep

m # Mass of the particle

R # Radius of the particle

Astroshaper.MutualOrbit — Type

mutable struct MutualOrbit{T1}

Mutual orbit parameters of a binary asteroid

Astroshaper.OrbitalElements — Type

Orbital elements

a : Semi-major axis

e : Eccentricity

I : Inclination

ω : Argument of periapsis

Ω : Longitude of ascending node

Φ : Mean anomaly (tₚ = - Φ / n)

tₚ : Periapsis passage time

μ : Standard gravitational parameter of the system (GM)

n : Mean motion : n = √(μ / a^3)

P : Orbital period : T = 2π / n

Time-variables

t : Time

u : Eccentric anomaly

ν : True anomaly

r : Position in the orbital plane frame (located at the focus of the ellipse)

v : Velocity in the orbital plane frame (located at the focus of the ellipse)

F☉ : Solar irradiation at the orbit [W/m²]

Astroshaper.ShapeModel — Type

ShapeModel{T1, T2, T3, T4, T5, T6, T7}

A polyhedral shape model of an asteroid.

Fields

num_node : Number of nodes

num_face : Number of faces

nodes : 1-D array of node positions

faces : 1-D array of vertex indices of faces

facets : 1-D array of surface facets (Facet)

AREA : Surface area

VOLUME : Volume

RADIUS_EQ : Equivalent radius of a sphere with the same volume

RADIUS_MAX : Maximum radius

RADIUS_MIN : Minimum radius

COF : Center-of-figure

MOI : Moment of inertia tensor

force : Thermal recoil force at body-fixed frame (Yarkovsky effect)

torque : Thermal recoil torque at body-fixed frame (YORP effect)

Astroshaper.SimpleParticle — Type

SimpleParticle{T1 <: AbstractVector, T2 <: Real} <: AbstractParticle

Paticle for N-body simulation (e.g. Euler or leapfrog integrator)

Fields

r # Position

v # Velocity

a # Acceleration

m # Mass of the particle

R # Radius of the particle

Astroshaper.SpinParams — Type

SpinParams{T1, T2}

Spin parameters of an asteroid

Fields

Spin pole @ Equatorial coordinate

α : Right ascension (RA)

δ : Declination (Dec)

Spin pole @ Ecliptic coordinate

λ : Ecliptic longitude

β : Ecliptic latitude

Other parameters

P : Spin period [sec]

ω : Angular velocity [rad/sec]

ŝ : Spin pole direction (normalized)

ε : Obliquity

γ : ernal equinox lon. from the direction of perihelion

t : Time

ϕ₀ : Initial spin phase

ϕ : Spin phase angle

Astroshaper.ThermoParams — Type

struct ThermoParams{COMMON_INT, COMMON_FLOAT, FACET_INT, FACET_FLOAT}

Fields

A_B : Bond albedo

A_TH : Albedo at thermal radiation wavelength

k : Thermal conductivity [W/m/K]

ρ : Material density [kg/m³]

Cp : Heat capacity [J/kg/K]

ϵ : Emissivity

t_bgn : Start time of the simulation, normalized by period P

t_end : End time of the simulation, normalized by period P

Δt : Non-dimensional timesteps, normalized by period P

Nt : Number of timesteps

z_max : Maximum depth for thermophysical simualtion, normalized by thermal skin depth l

Δz : Non-dimensional step in depth, normalized by thermal skin depth l

Nz : Number of depth steps

P : Cycle of thermal cycle (rotation period) [sec]

l : Thermal skin depth [m]

Γ : Thermal inertia [J ⋅ m⁻² ⋅ K⁻¹ ⋅ s⁻⁰⁵ (tiu)]

λ : Non-dimensional coefficient for heat diffusion equation

Parametic types

COMMON_INT : Integer-typed property common for all facets

COMMON_FLOAT : Float-typed property common for all facets

FACET_INT : Integer common for all facets or array giving an integer for each facet individually

FACET_FLOAT : Float common for all facets or array giving a float for each facet individually

Astroshaper.VisibleFacet — Type

struct VisibleFacet{T1, T2, T3}

Index of an interfacing facet and its view factor

Fields

id : Index of the interfacing facet

f : View factor from facet i to j

d : Distance from facet i to j

d̂ : Normal vector from facet i to j

Astroshaper.YORP_timescale — Method

YORP_timescale(P_start, P_end, ω̇) -> timescale

Caluculate the YORP time scale given a constant acceleration/deceleration

Parameters

P_start [hour]

P_end [hour]

ω̇ [deg/day/day]

Astroshaper.analyze_YORP — Method

analyze_YORP(df, spin, MOI)

Parameters

df : Dataframe of timestamp

spin : SpinParams

C : Moment of inertia

Astroshaper.autumnal_equinox_lon — Method

autumnal_equinox_lon(spin::SpinParams) -> γ_autum

Get a longitude of the autumnal equinox with respect to the perihelion

Astroshaper.body_to_orbit — Method

body_to_orbit(v, spin::SpinParams) -> v
body_to_orbit(v, γ, ε, ϕ)          -> v

Parameters

v : vector in the body-fixed frame

γ : longitude of vernal equinox direction of the body

ε : obliquity of the spin pole

ϕ : spin phase of the body

Return

v : vector in the orbital plane frame

Astroshaper.center_of_figure — Method

center_of_figure(facets) -> COF

Calculate center-of-figure position of a polyhedral

Astroshaper.drift_rate_ȧ — Method

a : Acceleration

u : Eccentric anomaly

orbit : OrbitalElements

Return

ȧ : Drift rate in semi-major axis

Astroshaper.eclipse_is_possible — Method

The secondary is within the critical angle to detect an eclipse event.

Astroshaper.energy_in — Method

energy_in(shape::ShapeModel, params::ThermoParams) -> E_in
energy_in(shape::ShapeModel, A_B::AbstractVector ) -> E_in
energy_in(shape::ShapeModel, A_B::Real           ) -> E_in
energy_in(facet::Facet,      A_B::Real           ) -> E_in

Input energy per second at a certain time [W]

Astroshaper.energy_io — Method

energy_io(shape::ShapeModel, params::ThermoParams) -> E_in, E_out, E_cons

Input and output energy per second at a certain time

Returns

E_in : Input energy per second at a certain time [W]

E_out : Output enegey per second at a certain time [W]

E_cons : Output-input energy ratio (E_out / E_in)

Astroshaper.energy_out — Method

energy_out(shape::ShapeModel, params::ThermoParams                   ) -> E_out
energy_out(shape::ShapeModel, ϵ::AbstractVector, A_TH::AbstractVector) -> E_out
energy_out(shape::ShapeModel, ϵ::Real,           A_TH::Real          ) -> E_out
energy_out(facet::Facet,      ϵ::Real,           A_TH::Real          ) -> E_out

Output enegey per second at a certain time [W]

Astroshaper.eq2ec — Method

eq2ec(α, δ) -> λ, β

Coordinate transform from the equatorial to ecliptic coordinate systems

Arguments

α: right ascension (RA)

δ: declination (Dec)

Returns

λ: ecliptic longitude

β: ecliptic latitude

Astroshaper.evaluate_Euler! — Method

⁺pᵢ : Particle i at next step

pᵢ : Particle i

pⱼ : Particle j

ϵ : Softening parameter

Astroshaper.findVisibleFacets! — Method

findVisibleFacets!(facets)

Find facets that is visible from each facet

Parameters

facets : Array of Facet

Astroshaper.findVisibleFacets! — Method

findVisibleFacets!(obs::Facet, facets)

Find facets that is visible from the facet where the observer is located.

Parameters

obs : Facet where the observer stands

facets : Array of Facet

Astroshaper.find_eclipse! — Method

find_eclipse!(shapes, sun_from_pri, sec_from_pri, R₂₁)

shapes

sun_from_pri

sec_from_pri : Position of the secondary relative to primary

R₂₁ : Rotation matrix from secondary to primary

Astroshaper.flux_total — Method

flux_total(shape::Shape, A_B::AbstractVector, A_TH::AbstractVector) -> F_total
flux_total(facet::Facet, A_B::Real,           A_TH::Real)           -> F_total

Arguments

shape : Shape model (Shape)

facet : Surface facet (Facet)

A_B : Bond albedo

A_TH : Albedo in thermal infrared wavelength

Total energy absorbed by the facet

Astroshaper.getBaryCenter — Method

getBaryCenter(ps) -> r_G, v_G, a_G

Position, velocity and acceleration of the system's barycenter

Astroshaper.getEdgeFactor — Method

getEdgeFactor(r_i, r_j) -> L_e

Calculate a edge factor, L_e

Arguments

r_i : start-point of the edge

r_j : end-point of the edge

pos : position where you want the gravity (if skipped, the origin is given)

Returns

L_e: edge factor

Astroshaper.getFaceAreas — Method

getFaceAreas(topology, points) -> faceareas

Get an array of areas of all faces

Arguments

topology : Topology instace of HalfEdges.jl

points : vertex-list of the polyhedron

Return

faceareas : array of areas of faces

Astroshaper.getFaceCenters — Method

getFaceCenters(topology, points) -> facecenters

Get an array of center-positions of all faces

Arguments

topology : Topology instace of HalfEdges.jl

points : vertex-list of the polyhedron

Return

facecenters : array of center-positions of faces

Astroshaper.getFaceDyads — Method

getFaceDyads(topology, points) -> facedyads

Get an array of dyads for all faces

Arguments

topology : Topology instace of HalfEdges.jl

points : vertex-list of the polyhedron

Return

facedyads : array of face dyads

Astroshaper.getFaceFactor — Method

getFaceFactor(r_i, r_j, r_k) -> ω_f

Calculate a face factor, ω_f

Arguments

r_i : 1st vertex of the face (counter-clockwise)

r_j : 2nd vertex of the face

r_k : 3rd vertex of the face

– pos : position where you want the gravity (if skipped, the origin is given)

Returns

ω_f: face factor

Astroshaper.getFaceNormals — Method

getFaceNormals(topology, points) -> facenormals

Get an array of normal vectors to all faces

Arguments

topology : Topology instace of HalfEdges.jl

points : vertex-list of the polyhedron

Return

facenormals : array of face normals

Astroshaper.getFaceVertices — Method

getFaceVertices(topology, points) -> r1s, r2s, r3s

Astroshaper.getHalfEdgeDyads — Method

getHalfEdgeDyads(topology, points) -> halfedgedyads

Get an array of dyads for all halfedges

Arguments

topology : Topology instace of HalfEdges.jl

points : vertex-list of the polyhedron

Return

halfedgedyads : array of halfedge dyads

Astroshaper.getHalfEdgeVertices — Method

getHalfEdgeVertices(topology, points) -> heads, tails

Astroshaper.getIlluminatedFacets — Method

getIlluminatedFacets(r̂☉, facets; ray_trace=true) -> ids

Arguments

r̂☉ : Normal vector to indicate solar direction

facets : Array of Facet

ray_trace : Option to turn on/off of ray-trace

Return

ids : Indices of illuminated facets

Astroshaper.getSolarCondition — Method

getSolarCondition(orbit, spin, time) -> Φ, r_sun

Get the solar irradition and the direction of the Sun

Parameters

orbit :

spin :

time : [sec]

Returns

F☉ : solar irradiation [W/m^2]

r̂☉ : solar direction in the body-fixed frame

Astroshaper.getSolarIrradiation — Method

getSolarIrradiation(distance) -> solar_irrad

Calculate the solar irradiation on the body

Parameter

rₕ : heliocentric distance of the body [m]

Return

F☉ : solar irradiation [W/m^2]

Astroshaper.get_r — Method

get_r(a, e, u) -> r

Get a body's positon based on Kepler's motion

Parameters

a : semi-major axis

e : eccentricity

u : eccentric anomaly

Return

r : position

(the coordinate system is in the orbital plane and its origin is the eclipse focus)

Astroshaper.get_rv — Method

get_rv(a, e, n, u) -> r, v

Get a body's positon and velocity based on Kepler's motion

Parameters

a : semi-major axis

e : eccentricity

n : mean motion

u : eccentric anomaly

Returns

r : position

v : velocity

(the coordinate system is in the orbital plane and its origin is the eclipse focus)

Astroshaper.get_v — Method

get_v(a, e, n, u) -> v

Get a body's velocity based on Kepler's motion

Parameters

a : semi-major axis

e : eccentricity

n : mean motion

u : eccentric anomaly

Return

v : velocity

(the coordinate system is in the orbital plane and its origin is the eclipse focus)

Astroshaper.getdyad — Method

getdyad(a, b) -> dyad

Make a dyad from two vectors. You can simply write: a * b'

Arguments

a, b: two vectors

Return

dyad: dyad

Astroshaper.getfacets — Method

Array of Facet, converted from arrays of nodes and faces of a shape model

Astroshaper.getvolume — Method

getvolume(facets) -> VOLUME

Calculate volume of a polyhedral

Astroshaper.inertia_to_orbit! — Method

Transform the inertial coordinate system to the orbital plane system

Parameters

v : vector to be transformed

ω : argument of periapsis

I : inclination

Ω : longitude of the ascneding node

Astroshaper.init_temps_zero! — Method

init_temps_zero!(shape::ShapeModel, params)
init_temps_zero!(shape::ShapeModel, Nz::AbstractVector)
init_temps_zero!(shape::ShapeModel, Nz::Integer)
init_temps_zero!(facet::Facet,      Nz::Integer)

Initialize temperature profile in depth on every facet. All elements are intialized as 0 K.

Astroshaper.intensity — Method

intensity(λ, T) -> I

Intensity of radiation at a wavelength λ and tempertature T according to the Planck function

Astroshaper.isAbove — Method

isAbove(A, B, C, D)             -> Bool
isAbove(facet::Facet, D)        -> Bool
isAbove(obs::Facet, tar::Facet) -> Bool

Determine if point D is above triangle facet ABC.

Astroshaper.isAboveHorizon — Method

isAboveHorizon(facet::Facet) -> Bool

Return if the facet is above its local horizon or not

Astroshaper.isBelow — Method

isBelow(A, B, C, D)             -> Bool
isBelow(facet::Facet, D)        -> Bool
isBelow(obs::Facet, tar::Facet) -> Bool

Determine if point D is below triangle facet ABC.

Astroshaper.isFace — Method

isFace(obs::Facet, tar::Facet) -> true/false

Determine if the two facets are facing each other

Astroshaper.isIlluminated — Method

isIlluminated(obs::AbstractVector, r̂☉, facets) -> Bool
isIlluminated(obs::Facet, r̂☉, facets)          -> Bool

Return if the observation point/facet is illuminated by the direct sunlight or not

Astroshaper.loadobj — Method

loadobj(shapepath::String; scale=1, message=true) -> nodes, faces

Astroshaper.mean_energy_cons_frac! — Method

mean_energy_cons_frac!(df, spin::SpinParams)
mean_energy_cons_frac!(df, P::Real)

Average energy conservation fraction over a rotational cycle

Astroshaper.moment_of_inertia — Method

moment_of_inertia(facets) -> MOI

Calculate moment of inertia tensor of a polyhedron

Astroshaper.orb_to_ref! — Method

Transform from the orbital plane system to the reference coordinate system

Parameters

r : vector to be transformed

ω : argument of periapsis

I : inclination

Ω : longitude of the ascneding node

Astroshaper.orbit_to_body — Method

orbit_to_body(v, spin::SpinParams) -> v
orbit_to_body(v, γ, ε, ϕ)          -> v

Parameters

v : vector in the orbital plane frame

γ : longitude of vernal equinox direction of the body

ε : obliquity of the spin pole

ϕ : spin phase of the body

Return

v : vector in the body-fixed frame

Astroshaper.orbit_to_inertia! — Method

Transform the orbital plane system to the inertial coordinate system

Parameters

v : vector to be transformed

ω : argument of periapsis

I : inclination

Ω : longitude of the ascneding node

Astroshaper.perturbnodes! — Method

Make a perturbed shape model

Astroshaper.prep_timestamp — Method

Astroshaper.pxform — Method

Arguments

from : Name of the frame to transform from

to : Name of the frame to transform to

ets : Epoch of the rotation matrix

Return

rotate : A rotation matrix

Astroshaper.raycast — Method

raycast(A, B, C, R)                           -> Bool
raycast(facet::Facet, R)                      -> Bool
raycast(facet::Facet, R, obs::AbstractVector) -> Bool
raycast(facet::Facet, R, obs::Facet)          -> Bool

Ray-triangle intersection detection. The ray R starts from the origin.

Astroshaper.ref_to_orb! — Method

Transform from the reference coordinate system to the orbital plane system

Parameters

r : vector to be transformed

ω : argument of periapsis

I : inclination

Ω : longitude of the ascneding node

Astroshaper.rotateX! — Method

Rotation transformation around X-axis Parameters: v : vector to be transformed θ : rotation angle

Astroshaper.rotateY! — Method

Rotation transformation around Y-axis Parameters: v : vector to be transformed θ : rotation angleon

Astroshaper.rotateZ! — Method

Rotation transformation around Z-axis Parameters: v : vector to be transformed θ : rotation angleon

Astroshaper.run_TPM! — Function

Astroshaper.run_TPM! — Method

shape : Shape model

et_range : Range of ephemeris times to run

sun : Sun's position in the body-fixed frame at epochs (Not normalized)

thermo_params : Thermophysical parametes

savepath : Path to save data file

save_range : Indices in et_range to be saved

Astroshaper.run_TPM! — Method

run_TPM!

Run TPM for a binary asteroid.

shapes

ephemerides

thermo_params

savepath

savevalues

Astroshaper.run_YORP — Method

Astroshaper.run_Yarkovsky — Method

Astroshaper.run_binary_TPM! — Method

run_binary_TPM!(binary, thermo_params)

Astroshaper.run_leapfrog — Method

run_leapfrog(ps, params_sim, filename)

Astroshaper.save_snapshot_txt — Method

save_snapshot(out::IOStream, t, ps)

Format of snapshot

t
n
m1 x1 y1 z1 vx1 vy1 vz1
m2 x2 y2 z2 vx2 vy2 vz2
:
:
mN xN yN zN vxN vyN vzN

Astroshaper.save_timestamp! — Method

Astroshaper.saveobj — Method

saveobj(path::String, nodes, faces)

Astroshaper.solid_angle — Method

getSolidAngle(facet::Facet, obs::AbstractVector) -> Ω

Solid angle of a triangular facet, equal to area of the corresponding spherical triangle

Astroshaper.solveKeplerEquation1 — Method

solveKeplerEquation1(e, n, tₚ, t) -> u_i

Solve Kepler's equation with the 1st order method

Parameters

e : eccentricity

n : mean motion

tₚ : periapsis passage time [sec]

t : time

Return

u_i : eccentric anomaly (i-th order solution)

Astroshaper.solveKeplerEquation2 — Method

solveKeplerEquation2(e, n, tₚ, t) -> u_i

Solve Kepler's equation with the 2nd order method (Newton's method)

Parameters

e : eccentricity

n : mean motion

tₚ : periapsis passage time [sec]

t : time

Return

u_i : eccentric anomaly (i-th order solution)

Astroshaper.spherical_excess — Method

spherical_excess(a, b, c) -> E

Area of a spherical triangle c.f. L'Huilier's Theorem

Arguments

a, b, c : side lengths of a spherical traiangle

Astroshaper.spin_normal — Method

Get a spin pole direction in the orbital plane frame

Paramters

λ : ecliptic longitude

β : ecliptic latitude

orbit : orbital elements

Return

ŝ : spin pole direction (normalized)

Astroshaper.spin_perp_unit1 — Method

spin_perp_unit1(spin::SpinParams) -> ê1

Get a unit vector ê⟂1

Astroshaper.spin_perp_unit2 — Method

spin_perp_unit2(spin::SpinParams) -> ê2

Get a unit vector ê⟂2

Astroshaper.spin_perp_units — Method

spin_perp_units(spin::SpinParams) -> ê1, ê2

Get a unit vector ê⟂1 and ê⟂2, perpendicular to spin pole

Astroshaper.spkpos — Method

Obtain a vector of position

Arguments

targ : Target body name

ets : AbstractVetor of observer epoch

ref : Reference frame of output position vector

abcorr : Aberration correction flag

obs : Observing body name

Astroshaper.spkpos_df — Method

Obtain a dataframe of ephemerides

Arguments

targ : Target body name

ets : AbstractVetor of observer epoch

ref : Reference frame of output position vector

abcorr : Aberration correction flag

obs : Observing body name

Astroshaper.step_heat_cond! — Method

step_heat_cond!(shape::ShapeModel, λ::AbstractVector)
step_heat_cond!(shape::ShapeModel, λ::Real)
step_heat_cond!(facet::Facet,      λ::Real)
step_heat_cond!(Tⱼ::AbstractVector, Tⱼ₊₁::AbstractVector, λ::Real)

Calculate temperature profile at the next step and update Facet.temps

Arguments

facet : Surface facet (Facet)

λ : Coefficient of heat conduction equation

Tⱼ : Temperatures

Tⱼ₊₁ : Temperatures at the next timestep

Astroshaper.sum_force_torque! — Method

sum_force_torque!(shape::Shape)

Integrate the force and torque over the global surface

Astroshaper.thermal_inertia — Method

thermal_inertia(params)   -> Γ
thermal_inertia(k, ρ, Cp) -> Γ

Arguments

k : Thermal conductivity [W/m/K]

ρ : Material density [kg/m³]

Cp : Heat capacity [J/kg/K]

Return

Γ : Thermal inertia [J ⋅ m⁻² ⋅ K⁻¹ ⋅ s⁻⁰⁵ (tiu)]

Astroshaper.thermal_skin_depth — Method

thermal_skin_depth(params)      -> l_2π
thermal_skin_depth(P, k, ρ, Cp) -> l_2π

Arguments

P : Cycle of thermal cycle [sec]

k : Thermal conductivity [W/m/K]

ρ : Material density [kg/m³]

Cp : Heat capacity [J/kg/K]

Return

l_2π : Thermal skin depth [m]

Astroshaper.u2ν — Method

u2ν(u, e) -> ν

Calculate true anomaly ν from eccentric anomaly u and eccentricity e

Parameters

u : Eccentric anomaly

e : Eccentricity

Return

ν : True anomaly

Astroshaper.update! — Method

update_binary!(binary, t)

Astroshaper.update_bottom_temp! — Method

Update bottom temperature under boundary condition of insulation

Astroshaper.update_flux! — Method

update_flux!(shape, F☉, r̂☉, thermo_params)

Update energy flux to every facet by solar radiation, scattering, and re-absorption of radiation

Astroshaper.update_flux_rad_single! — Method

update_flux_rad_single!(shape, params)
update_flux_rad_single!(shape, ϵ::Real)
update_flux_rad_single!(shape, ϵ::AbstractVector)

Single radiation-reabsorption is considered, assuming albedo is close to zero at thermal infrared wavelength.

Astroshaper.update_flux_scat_single! — Method

update_flux_scat_single!(shape, params)
update_flux_scat_single!(shape, A_B::Real)
update_flux_scat_single!(shape, A_B::AbstractVector)

Single scattering of sunlight is considered.

Astroshaper.update_flux_sun! — Method

update_flux_sun!(shape, r☉::AbstractVector)

Update solar radiation flux on every facet of a shape model.

shape : Shape model

r☉ : Position of the sun in the body-fixed frame, which is not normalized.

Astroshaper.update_flux_sun! — Method

update_flux_sun!(shape, F☉, r̂☉)

Update solar radiation flux on every facet of a shape model.

shape : Shape model

F☉ : Solar radiation flux

r̂☉ : Unit vector indicating the direction of the sun in the body-fixed frame

Astroshaper.update_force! — Method

update_force!(shape::ShapeModel, params)
update_force!(shape::ShapeModel, A_B::Real,           ϵ::Real)
update_force!(shape::ShapeModel, A_B::AbstractVector, ϵ::Real)
update_force!(shape::ShapeModel, A_B::Real,           ϵ::AbstractVector)
update_force!(shape::ShapeModel, A_B::AbstractVector, ϵ::AbstractVector)
update_force!(facet::Facet,      A_B::Real,           ϵ::Real)

Update photon recoil force on every facet (dfᵢ)

Astroshaper.update_force_Rubincam! — Method

update_force_Rubincam!(shape, params_thermo)

Update photon recoil force on every facet (df) based on Rubincam (2000) approximation

Astroshaper.update_orbit! — Method

update_orbit!(orbit::OrbitalElements, t)

Astroshaper.update_spin! — Method

update_spin!(spin::SpinParams, t)

Astroshaper.update_surf_temp! — Method

update_surf_temp!(shape::ShapeModel, A_B, A_TH, k, l, Δz, ϵ)
update_surf_temp!(shape::ShapeModel, A_B::Real, A_TH::Real, k::Real, l::Real, Δz::Real, ϵ::Real)
update_surf_temp!(facet::Facet,      A_B::Real, A_TH::Real, k::Real, l::Real, Δz::Real, ϵ::Real)
update_surf_temp!(T::AbstractVector, F_total::Real, k::Real, l::Real, Δz::Real, ϵ::Real)

Update surface temperature under radiative boundary condition using Newton's method

Arguments

facet : surface facet (Facet)

A_B : Bond albedo

A_TH : Albedo in thermal infrared wavelength

k : Thermal conductivity

l : Thermal skin depth

Δz : Step width in depth direction (normalized by thermal skin depth l)

ϵ : Emissivity

T : 1-D array of temperatures

F_total : Total energy absorbed by the facet

In the normalized equation of the surface boundary condition, the coefficient Γ / √(4π * P) is equivalent for k / l, where Γ is the thermal inertia and P the rotation period.

Astroshaper.update_temps! — Method

update_temps!(shape::ShapeModel, params::ThermoParams)
update_temps!(shape::ShapeModel, λ, A_B, A_TH, k, l, Δz, ϵ)
update_temps!(facet::Facet, λ::Real, A_B::Real, A_TH::Real, k::Real, l::Real, Δz::Real, ϵ::Real)

Update temerature profie (Facet.temps) based on 1-D heat diffusion

Astroshaper.vernal_equinox_lon — Method

vernal_equinox_lon(spin::SpinParams)        -> γ
vernal_equinox_lon(ŝ::AbstractVector) -> γ

Get a longitude of the vernal equinox with respect to the perihelion

Astroshaper.view_factor — Method

view_factor(i::Facet, j::Facet) -> fᵢⱼ, dᵢⱼ, d̂ᵢⱼ

View factor from facet i to j, assuming Lambertian emission